mbsprint: improve printing output when it has invalid UTF data
Reset the decode state when mbtowc returns -1. The OpenBSD mbtowc(3)
man page says: "If a call to mbtowc() resulted in an undefined internal
state, mbtowc() must be called with s set to NULL to reset the internal
state before it can safely be used again."

Print the UTF replacement character (codepoint 0xfffd) for the invalid
codepoint or incomplete sequence and continue printing the line
(instead of stopping).

Remove the 0 return code as it can't happen because we're already
checking the string length in the loop.